+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added actual types for `@znotfireman/ohmyprvd`
+- Added missing error messages
+
 ## 0.1.0-alpha
 
 - Initial release